ADD: address all PR comments
