+++ src/main/scala/trace/TraceCoreInterface.scala
@@ -1,7 +1,7 @@
 // See LICENSE.Berkeley for license details.
 // See LICENSE.SiFive for license details.
 
-package freechips.rocketchip.util
+package freechips.rocketchip.trace
 
 import chisel3._
 
@@ -25,12 +25,13 @@ object TraceItype extends ChiselEnum {
   val ITInJump    = Value(15.U)
 }
 
-class TraceCoreParams (
-  val nGroups: Int = 1,
-  val iretireWidth: Int = 1,
-  val xlen: Int = 64,
-  val iaddrWidth: Int = 64
+case class TraceCoreParams (
+  nGroups: Int = 1,
+  iretireWidth: Int = 1,
+  xlen: Int = 32,
+  iaddrWidth: Int = 32
 )
+
 
 class TraceCoreGroup (val params: TraceCoreParams) extends Bundle {
   val iretire = UInt(params.iretireWidth.W)
