+++ src/main/scala/trace/TraceEncoder.scala
@@ -1,20 +1,20 @@
 // See LICENSE.Berkeley for license details.
 // See LICENSE.SiFive for license details.
 
-package freechips.rocketchip.util
+package freechips.rocketchip.trace
 
 import chisel3._
 import chisel3.util._
 import scala.math.min
 
 
-class TraceEncoderParams(
-  val coreParams: TraceCoreParams,
-  val bufferDepth: Int,
-  val encoderBaseAddr: BigInt,
-  val sinkDMABaseAddr: BigInt,
-  val useSinkPrint: Boolean,
-  val useSinkDMA: Boolean
+case class TraceEncoderParams(
+  coreParams: TraceCoreParams,
+  bufferDepth: Int,
+  encoderBaseAddr: BigInt,
+  sinkDMABaseAddr: BigInt,
+  useSinkPrint: Boolean,
+  useSinkDMA: Boolean
 )
 
 object FullHeaderType extends ChiselEnum {
