+++ src/main/scala/trace/TraceSinkDMA.scala
@@ -1,49 +1,13 @@
-package freechips.rocketchip.util
+package freechips.rocketchip.trace
 
 import chisel3._
 import chisel3.util._
-import chisel3.experimental.StringParam
 import freechips.rocketchip.tilelink._
 import freechips.rocketchip.diplomacy._
 import org.chipsalliance.cde.config.{Parameters, Field, Config}
 import freechips.rocketchip.subsystem._
 import freechips.rocketchip.prci._
 import freechips.rocketchip.regmapper.{RegField, RegFieldDesc}
-
-trait HasTraceSinkIO {
-  val io = IO(new Bundle {
-    val in = Flipped(Decoupled(UInt(8.W)))
-  })
-}
-
-class TraceSinkPrint(bb_name: String)(implicit p: Parameters) extends LazyModule {
-  override lazy val module = new TraceSinkPrintImpl(this)
-  val blackbox_name = bb_name
-  class TraceSinkPrintImpl(outer: TraceSinkPrint) extends LazyModuleImp(outer) with HasTraceSinkIO {
-    withClockAndReset(clock, reset) {
-      io.in.ready := true.B
-      val byte_printer = Module(new BytePrinter(outer.blackbox_name))
-      byte_printer.io.clk := clock
-      byte_printer.io.reset := reset
-      byte_printer.io.in_valid := io.in.valid
-      byte_printer.io.in_byte := io.in.bits
-    }
-  }
-}
-
-class BytePrinter(name: String) extends BlackBox(
-  Map(
-    "FILE_NAME" -> StringParam(s"tacit_${name}_trace.out")
-  )
-) with HasBlackBoxResource {
-  val io = IO(new Bundle {
-    val clk = Input(Clock())
-    val reset = Input(Reset())
-    val in_valid = Input(Bool())
-    val in_byte = Input(UInt(8.W))
-  })
-  addResource("/vsrc/BytePrinter.v")
-}
 
 class TraceSinkDMA(addr: BigInt, beatBytes: Int)(implicit p: Parameters) extends LazyModule {
   val node = TLClientNode(Seq(TLMasterPortParameters.v1(Seq(TLClientParameters(
@@ -59,7 +23,7 @@ class TraceSinkDMA(addr: BigInt, beatBytes: Int)(implicit p: Parameters) extends
   lazy val module = new TraceSinkDMAImpl(this)
   class TraceSinkDMAImpl(outer: TraceSinkDMA) extends LazyModuleImp(outer) with HasTraceSinkIO {
     val fifo = Module(new Queue(UInt(8.W), 32))
-    fifo.io.enq <> io.in
+    fifo.io.enq <> io.trace_in
     val (mem, edge) = outer.node.out(0)
     val addrBits = edge.bundle.addressBits
     val busWidth = edge.bundle.dataBits
